pyproject.toml
[tool.poetry]
name = "thesis-research"
version = "0.1.0"
description = ""
authors = ["mario-dg <mario.dagraca@gmx.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.10.12"
absl-py = "2.1.0"
accelerate = "0.27.2"
aiofiles = "22.1.0"
aiohttp = "3.9.1"
aiosignal = "1.3.1"
aiosqlite = "0.19.0"
annotated-types = "0.6.0"
antlr4-python3-runtime = "4.9.3"
anyio = "4.2.0"
appdirs = "1.4.4"
argon2-cffi = "23.1.0"
argon2-cffi-bindings = "21.2.0"
arrow = "1.3.0"
asttokens = "2.4.1"
async-lru = "2.0.4"
async-timeout = "4.0.3"
attrs = "23.2.0"
babel = "2.14.0"
backoff = "2.2.1"
beautifulsoup4 = "4.12.3"
bleach = "6.1.0"
boto3 = "1.34.25"
botocore = "1.34.25"
cachetools = "5.3.2"
certifi = "2023.11.17"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
comm = "0.2.1"
conda-pack = "0.7.1"
contourpy = "1.2.0"
cycler = "0.12.1"
datasets = "2.16.1"
debugpy = "1.8.0"
decorator = "5.1.1"
defusedxml = "0.7.1"
diffusers = "0.26.3"
dill = "0.3.7"
docker-pycreds = "0.4.0"
entrypoints = "0.4"
exceptiongroup = "1.2.0"
executing = "2.0.1"
fastapi = "0.109.0"
fastjsonschema = "2.19.1"
filelock = "3.13.1"
fire = "0.5.0"
fonttools = "4.47.2"
fqdn = "1.5.1"
frozenlist = "1.4.1"
fsspec = "2023.10.0"
ftfy = "6.1.3"
gitdb = "4.0.11"
gitpython = "3.1.41"
google-auth = "2.26.2"
google-auth-oauthlib = "1.2.0"
grpcio = "1.60.0"
h11 = "0.14.0"
huggingface-hub = "0.20.3"
hydra-core = "1.3.2"
idna = "3.6"
importlib-metadata = "7.0.1"
ipykernel = "6.26.0"
ipython = "8.17.2"
ipython-genutils = "0.2.0"
ipywidgets = "8.1.1"
isoduration = "20.11.0"
jedi = "0.19.1"
jinja2 = "3.1.3"
jmespath = "1.0.1"
joblib = "1.3.2"
json5 = "0.9.14"
jsonpointer = "2.4"
jsonschema = "4.21.1"
jsonschema-specifications = "2023.12.1"
kiwisolver = "1.4.5"
lightning = "2.1.2"
lightning-cloud = "0.5.61"
lightning-utilities = "0.10.1"
lightning-sdk = "0.0.14a0"
markdown = "3.5.2"
markdown-it-py = "3.0.0"
markupsafe = "2.1.4"
matplotlib = "3.8.2"
matplotlib-inline = "0.1.6"
mdurl = "0.1.2"
mistune = "3.0.2"
mpmath = "1.3.0"
multidict = "6.0.4"
multiprocess = "0.70.15"
nbclassic = "1.0.0"
nbclient = "0.9.0"
nbconvert = "7.14.2"
nbformat = "5.9.2"
nest-asyncio = "1.6.0"
networkx = "3.2.1"
notebook = "6.5.6"
notebook-shim = "0.2.3"
numpy = "1.26.2"
oauthlib = "3.2.2"
omegaconf = "2.3.0"
overrides = "7.6.0"
packaging = "23.2"
pandas = "2.1.4"
pandocfilters = "1.5.1"
parso = "0.8.3"
pexpect = "4.9.0"
pillow = "10.2.0"
platformdirs = "4.1.0"
prometheus-client = "0.19.0"
prompt-toolkit = "3.0.43"
protobuf = "4.23.4"
psutil = "5.9.8"
ptyprocess = "0.7.0"
pure-eval = "0.2.2"
pyarrow = "15.0.0"
pyarrow-hotfix = "0.6"
pyasn1 = "0.5.1"
pyasn1-modules = "0.3.0"
pycparser = "2.21"
pydantic = "2.5.3"
pydantic-core = "2.14.6"
pygments = "2.17.2"
pyjwt = "2.8.0"
pyparsing = "3.1.1"
python-dateutil = "2.8.2"
python-json-logger = "2.0.7"
python-multipart = "0.0.6"
pytorch-lightning = "2.1.2"
pytz = "2023.3.post1"
pyyaml = "6.0.1"
pyzmq = "24.0.1"
referencing = "0.32.1"
regex = "2023.12.25"
requests = "2.31.0"
requests-oauthlib = "1.3.1"
rfc3339-validator = "0.1.4"
rfc3986-validator = "0.1.1"
rich = "13.7.0"
rpds-py = "0.17.1"
rsa = "4.9"
s3transfer = "0.10.0"
safetensors = "0.4.2"
scikit-learn = "1.3.2"
scipy = "1.12.0"
send2trash = "1.8.2"
sentry-sdk = "1.39.2"
setproctitle = "1.3.3"
simple-term-menu = "1.6.4"
six = "1.16.0"
smmap = "5.0.1"
sniffio = "1.3.0"
soupsieve = "2.5"
stack-data = "0.6.3"
starlette = "0.35.1"
sympy = "1.12"
tensorboard = "2.15.1"
tensorboard-data-server = "0.7.2"
termcolor = "2.4.0"
terminado = "0.18.0"
threadpoolctl = "3.2.0"
tinycss2 = "1.2.1"
tokenizers = "0.15.2"
tomli = "2.0.1"
torch-ema = "0.3"
torch-fidelity = "0.3.0"
torchmetrics = "1.2.0"
tornado = "6.4"
tqdm = "4.66.1"
traitlets = "5.14.1"
transformers = "4.37.2"
types-python-dateutil = "2.8.19.20240106"
typing-extensions = "4.9.0"
tzdata = "2023.4"
uri-template = "1.3.0"
urllib3 = "2.0.7"
uvicorn = "0.27.0"
wandb = "0.16.6"
wcwidth = "0.2.13"
webcolors = "1.13"
webencodings = "0.5.1"
websocket-client = "1.7.0"
werkzeug = "3.0.1"
widgetsnbextension = "4.0.9"
xxhash = "3.4.1"
y-py = "0.6.2"
yarl = "1.9.4"
ypy-websocket = "0.8.4"
zipp = "3.17.0"
torch = "2.1.2"
torchvision = "0.16.2"
triton = "2.1.0"


[build-system]
requires = ["poetry-core"]